@@ -1,7 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatFormFieldModule } from '@angular/material/form-field';
-import { ReactiveFormsModule } from '@angular/forms'; 
-import { AuthService } from '../../services/auth.service'; 
+import { ReactiveFormsModule } from '@angular/forms';
+import { StoreModule } from '@ngrx/store'; // Importa StoreModule
+import { AuthService } from '../../services/auth.service';
 import { RegisterComponent } from './register.component';
 import { HttpClientModule } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
@@ -15,7 +16,15 @@ describe('RegisterComponent', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       declarations: [RegisterComponent],
-      imports: [ReactiveFormsModule, HttpClientModule, MatFormFieldModule, HttpClientTestingModule, MatInputModule, BrowserAnimationsModule], 
+      imports: [
+        ReactiveFormsModule,
+        HttpClientModule,
+        MatFormFieldModule,
+        HttpClientTestingModule,
+        MatInputModule,
+        BrowserAnimationsModule,
+        StoreModule.forRoot({}), 
+      ],
       providers: [AuthService],
     });
     fixture = TestBed.createComponent(RegisterComponent);
