@@ -3,8 +3,9 @@ import { HttpClientModule } from '@angular/common/http';
 import { StudentsComponent } from './students.component';
 import { StudentService } from './services/student.service'; 
 import { MatDialogModule } from '@angular/material/dialog';
-import { StudentsTableComponent } from './components/students-table/students-table.component';  // Import the StudentsTableComponent
+import { StudentsTableComponent } from './components/students-table/students-table.component';  
 import { MatCardModule } from '@angular/material/card';
+import { StoreModule } from '@ngrx/store'; 
 
 describe('StudentsComponent', () => {
   let component: StudentsComponent;
@@ -12,9 +13,14 @@ describe('StudentsComponent', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [StudentsComponent, StudentsTableComponent],  // Include StudentsTableComponent in the declarations
-      imports: [HttpClientModule, MatDialogModule,MatCardModule],
-      providers: [StudentService],  // Provide your services here
+      declarations: [StudentsComponent, StudentsTableComponent],
+      imports: [
+        HttpClientModule,
+        MatDialogModule,
+        MatCardModule,
+        StoreModule.forRoot({}) 
+      ],
+      providers: [StudentService],
     });
     fixture = TestBed.createComponent(StudentsComponent);
     component = fixture.componentInstance;
