@@ -1,9 +1,14 @@
 export interface User {
-    id: string;
+    id: number;
     name: string;
     lastname: string;
     email: string;
     password: string;
     token: string; // Puedes ajustar el tipo de token según la estructura real
-    createdAt: string ;
+    role: string;
   }
+
+export interface UserPayload {
+  courseId: number | null;
+  userId: number | null;
+}
