@@ -1,28 +1,36 @@
-import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
-import { BehaviorSubject, Observable , map} from 'rxjs';
+import { BehaviorSubject, Observable, map } from 'rxjs';
 import { User } from '../models/user';
 import { LoginPayload } from '../models/login';
 import { RegisterPayload } from '../models/register';
+import { Store } from '@ngrx/store';
+import { AuthActions } from '../store/auth.actions';
+import { selectAuthUser } from '../store/auth.selector';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AuthService {
-
   private _authUser$ = new BehaviorSubject<User | null>(null);
 
-  public authUser$ = this._authUser$.asObservable();
+  public authUser$ = this.store.select(selectAuthUser);
 
-  private apiUrl = "https://654bf2e15b38a59f28eff3b9.mockapi.io/api/v1/";
+  private apiUrl = 'https://654bf2e15b38a59f28eff3b9.mockapi.io/api/v1/';
 
-  constructor(private httpClient: HttpClient, private router: Router) {}
+  constructor(
+    private httpClient: HttpClient,
+    private router: Router,
+    private store: Store
+  ) {}
+
+  private handleAuthUser(authUser: User): void {
+    this.store.dispatch(AuthActions.setAuthUser({ data: authUser }));
+    localStorage.setItem('token', authUser.token);
+  }
 
   login(payload: LoginPayload): void {
-    // const headers = new HttpHeaders({
-    //   token: localStorage.getItem('token') || 'NO HAY TOKEN',
-    // });
     this.httpClient
       .get<User[]>(
         `${this.apiUrl}/users?email=${payload.email}&password=${payload.password}`
@@ -30,24 +38,23 @@ export class AuthService {
       .subscribe({
         next: (response) => {
           if (!response.length) {
-            alert('Usuario o contrasena invalidos');
+            alert('Usuario o contraseña inválidos');
           } else {
             const authUser = response[0];
-            if(authUser.password == payload.password){
-              this._authUser$.next(authUser);
-              localStorage.setItem('token', authUser.token);
+            if (authUser.password == payload.password) {
+              this.handleAuthUser(authUser);
               this.router.navigate(['/dashboard/home']);
-            }else{
-              alert('Usuario o contrasena invalidos');
+            } else {
+              alert('Usuario o contraseña inválidos');
             }
-            
           }
         },
         error: (err) => {
-          alert('Error de conexion');
+          alert('Error de conexión');
         },
       });
   }
+
   register(payload: RegisterPayload): void {
     this.httpClient
       .post<User>(
@@ -57,12 +64,13 @@ export class AuthService {
           password: payload.password,
           name: payload.name,
           lastname: payload.lastname,
-          token: payload.token
+          token: payload.token,
+          role: payload.role
         }
       )
       .subscribe({
         next: (authUser) => {
-          this._authUser$.next(authUser);
+          this.handleAuthUser(authUser);
           alert('Usuario creado');
           this.router.navigate(['/auth/login']);
         },
@@ -71,17 +79,6 @@ export class AuthService {
         },
       });
   }
-  
-  // "name": "Bernadette",
-  // "lastname": "Johns",
-  // "email": "Ned31@gmail.com",
-  // "password": "1gehMIgQB5SLvXO",
-  // "token": "tvVDmdIuraGJfrH",
-  // "createdAt": "2023-11-08T11:11:21.828Z",
-  // "updatedAt": 1699479118,
-  // "deletedAt": 1699479118,
-  // "id": "1"
-  
 
   verifyToken(): Observable<boolean> {
     return this.httpClient
@@ -94,8 +91,7 @@ export class AuthService {
             return false;
           } else {
             const authUser = users[0];
-            this._authUser$.next(authUser);
-            localStorage.setItem('token', authUser.token);
+            this.handleAuthUser(authUser);
             return true;
           }
         })
