@@ -1,5 +1,8 @@
-import { Component } from '@angular/core';
+import { Component, EventEmitter, Output } from '@angular/core';
 import { Router } from '@angular/router';
+import { Observable, map, of } from 'rxjs';
+import { User } from 'src/app/auth/models/user';
+import { AuthService } from 'src/app/auth/services/auth.service';
 
 @Component({
   selector: 'app-sidenav',
@@ -7,15 +10,28 @@ import { Router } from '@angular/router';
   styleUrls: ['./sidenav.component.scss']
 })
 export class SidenavComponent {
-  constructor(private router: Router) {
+
+  public authUser$: Observable<User | null>;
+
+  constructor(private router: Router, private authService: AuthService) {
     // Inicialmente, establece isActive según la ruta actual
     this.isActive = this.router.isActive('/dashboard/home', true);
+    this.authUser$ = this.authService.authUser$;
   }
 
+  get role$(): Observable<string | undefined> {
+    return this.authUser$?.pipe(
+      map((user) => user?.role)
+    ) || of(undefined);
+  }
+  
   isActive: boolean;
 
   // Función para cambiar el estado activo
   setActive(isActive: boolean) {
     this.isActive = isActive;
   }
+
+  @Output()
+  logOut = new EventEmitter();
 }
