@@ -3,6 +3,6 @@ export interface Course {
     nombre: string;
     fecha_inicio: string;
     fecha_fin: string;
-    clases: [] | null;
+    clases: number[] | null;
     descripcion: string;
   }
