@@ -1,8 +1,11 @@
 import { Component,Inject } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatDialogRef, MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
 import { Student } from 'src/app/dashboard/models/student';
 import { StudentService } from '../../services/student.service';
+import { CourseService } from '../../../courses/services/course.service';
+import { NavigationEnd, Router } from '@angular/router';
+import { filter } from 'rxjs';
 
 @Component({
   selector: 'app-students-dialog',
@@ -11,11 +14,15 @@ import { StudentService } from '../../services/student.service';
 })
 export class StudentsDialogComponent {
 
+  lastPartOfCourses: any[] = [];
+  selectedCourseIds: number[] = []; 
+
   nombreControl = new FormControl();
   apellidoControl = new FormControl();
   edadControl = new FormControl();
   generoControl = new FormControl();
   emailControl = new FormControl();
+  cursosControl = new FormControl();
 
   studentForm = new FormGroup({
     nombre: this.nombreControl,
@@ -23,12 +30,16 @@ export class StudentsDialogComponent {
     edad: this.edadControl,
     genero: this.generoControl,
     email: this.emailControl,
+    cursos: this.cursosControl
   });
 
 constructor(
   private fb: FormBuilder,
   private matDialogRef: MatDialogRef<StudentsDialogComponent>,
+  private matDialog: MatDialog,
   private studentService: StudentService,
+  private courseService: CourseService,
+  private router: Router,
 
   // RECIBO LA DATA (Student)
   @Inject(MAT_DIALOG_DATA) public data:{student:number, tipo:string}
@@ -43,6 +54,26 @@ constructor(
     })
   }
 
+  if(data.tipo === 'addcourse'){
+    this.courseService.getCourses().subscribe({
+      next: (c) =>{
+        if (c) {
+          this.lastPartOfCourses = c;
+        }
+      }
+    })
+  }
+
+  this.router.events.pipe(
+    filter((event) => event instanceof NavigationEnd)
+  ).subscribe(() => {
+    this.matDialog.closeAll();
+  });
+
+}
+
+goToCourses(): void {
+  this.router.navigate(['/dashboard/courses']);
 }
 
 public get isEditing(): boolean {
@@ -57,6 +88,15 @@ public get isAdding(): boolean {
   return this.data.tipo === 'add';
 }
 
+public get isDetails(): boolean {
+  return this.data.tipo === 'details';
+}
+
+public get isInscription(): boolean {
+  return this.data.tipo === 'addcourse';
+}
+
+
 
 
 
@@ -64,9 +104,14 @@ onSubmit(): void {
   if (this.studentForm.invalid) {
     return this.studentForm.markAllAsTouched();
   } else {
-    this.matDialogRef.close(this.studentForm.value);
+    const result = {
+      ...this.studentForm.value,
+      cursos: this.selectedCourseIds  // Agrega la propiedad 'cursos' al objeto result
+    };
+    this.matDialogRef.close(result);
   }
 }
+
 
 
 }
