@@ -3,18 +3,36 @@ import { Observable, map, take } from 'rxjs';
 import { StudentService } from './services/student.service';
 import { MatDialog } from '@angular/material/dialog';
 import { StudentsDialogComponent } from './components/students-dialog/students-dialog.component';
+import { AuthService } from 'src/app/auth/services/auth.service';
+import { User } from 'src/app/auth/models/user';
 
 @Component({
   selector: 'app-students',
   templateUrl: './students.component.html',
   styleUrls: ['./students.component.scss']
 })
-export class StudentsComponent {
+export class StudentsComponent  {
 
+  public authUser$: Observable<User | null>;
   students$: Observable<any>;
+  role$: Observable<string | undefined>;
 
-  constructor(private studentService: StudentService, private matDialog: MatDialog){
+  constructor(private studentService: StudentService, private matDialog: MatDialog, private authService: AuthService){
     this.students$ = this.studentService.getStudents();
+    this.authUser$ = this.authService.authUser$;
+    this.role$ = this.authUser$.pipe(map((user) => user?.role));
+    
+  }
+
+ 
+
+  onDetails(studentId:number):void{
+    const tipo = 'details';
+  
+    this.matDialog
+      .open(StudentsDialogComponent, {
+        data: { student: studentId, tipo: tipo }
+      })
   }
 
   addStudent(): void {
@@ -53,6 +71,31 @@ export class StudentsComponent {
         },
       });
   }
+
+  onAddCourseStudent(studentId: number): void {
+    let tipo = 'addcourse';
+    this.matDialog
+      .open(StudentsDialogComponent, {
+        data: { tipo }
+      })
+      .afterClosed()
+      .subscribe({
+        next: (result) => {
+          
+          if (result && result.cursos && result.cursos.length > 0) {
+            const cursosSeleccionados = result.cursos.map((cursoId: any) => Number(cursoId));
+            this.studentService.addCoursesToStudent(studentId, cursosSeleccionados).subscribe(
+              (updatedStudent) => {
+                this.students$ = this.studentService.getStudents();
+              },
+              (error) => {
+                console.error('Error al agregar cursos al estudiante', error);
+              }
+            );
+          }
+        },
+      });
+  }
   
   
 
@@ -67,15 +110,11 @@ export class StudentsComponent {
       .subscribe({
         next: (result) => {
           if (result) {
-            // Llama al servicio para realizar la eliminación
             this.studentService.deleteStudent(studentId).subscribe(
               (updatedStudents) => {
-                // Actualiza la tabla con la nueva lista de estudiantes después de la eliminación.
                 this.students$ = this.studentService.getStudents();
-                console.log('Estudiante eliminado:', result);
               },
               (error) => {
-                // Maneja el error si es necesario.
                 console.error('Error al eliminar estudiante', error);
               }
             );
@@ -101,7 +140,6 @@ export class StudentsComponent {
                 this.students$ = this.studentService.getStudents();
               },
               (error) => {
-                // Maneja el error si es necesario.
                 console.error('Error al editar estudiante', error);
               }
             );
