@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { MatTableModule } from '@angular/material/table';
 import { MatIconModule } from '@angular/material/icon';
 import { MatDialogModule } from '@angular/material/dialog';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatCardModule } from '@angular/material/card';
@@ -13,6 +13,9 @@ import { FormErrorsPipe } from './pipes/pipes/form-errors.pipe';
 import {MatButtonModule} from '@angular/material/button';
 import {MatToolbarModule} from '@angular/material/toolbar';
 import { FullnamePipe } from './pipes/pipes/fullname.pipe';
+import { MatDividerModule } from '@angular/material/divider';
+import { MatListModule } from '@angular/material/list';
+import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 
 
 @NgModule({
@@ -28,6 +31,7 @@ import { FullnamePipe } from './pipes/pipes/fullname.pipe';
     MatIconModule,
     MatDialogModule,
     ReactiveFormsModule,
+    FormsModule,
     MatFormFieldModule,
     MatInputModule,
     MatCardModule,
@@ -36,7 +40,10 @@ import { FullnamePipe } from './pipes/pipes/fullname.pipe';
     FormErrorsPipe,
     MatToolbarModule,
     MatButtonModule,
-    FullnamePipe
+    FullnamePipe,
+    MatDividerModule,
+    MatListModule,
+    MatProgressSpinnerModule
   ]
 })
 export class SharedModule { }
